norme

+++ includes/fractol.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   fractol.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gcollett <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2017/09/19 02:13:11 by gcollett          #+#    #+#             */
+/*   Updated: 2017/09/19 02:49:00 by gcollett         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FRACTOL_H
 # define FRACTOL_H
 
@@ -8,55 +20,55 @@
 # define WIDTH 750
 # define HEIGHT 750
 
-typedef struct	s_point
+typedef struct		s_point
 {
-	double	x;
-	double	y;
-}								t_point;
+	double			x;
+	double			y;
+}					t_point;
 
-typedef struct	s_modif
+typedef struct		s_modif
 {
-	int 		current;
+	int				current;
 	long			back;
-	int			distance;
-	double		x_base;
-	double		y_base;
-	double		zoom;
-}				t_modif;
+	int				distance;
+	double			x_base;
+	double			y_base;
+	double			zoom;
+}					t_modif;
 
-typedef struct	s_img
+typedef struct		s_img
 {
-	int			sl;
-	int			bpp;
-	int			end;
-	int			*image;
-	int			choice;
-	int			iteration_max;
-	void		*img_addr;
-	double		x[2];
-	double		y[2];
-	double		x_scale;
-	double		y_scale;
-  struct  s_img   *next;
-  struct  s_img   *prev;
-	t_point	inc;
-	t_modif	*modif;
-}				t_img;
+	int				sl;
+	int				bpp;
+	int				end;
+	int				*image;
+	int				choice;
+	int				iteration_max;
+	void			*img_addr;
+	float			x[2];
+	float			y[2];
+	float			x_scale;
+	float			y_scale;
+	struct s_img	*next;
+	struct s_img	*prev;
+	t_point			inc;
+	t_modif			*modif;
+}					t_img;
 
-typedef struct	s_map
+typedef struct		s_map
 {
-	void		*mlx;
-	void		*win;
-	int			**matrice;
-	int			*len_line;
-	int			len;
-	t_img		*img;
-}				t_map;
+	void			*mlx;
+	void			*win;
+	int				**matrice;
+	int				*len_line;
+	int				len;
+	t_img			*img;
+}					t_map;
 
-void	re_trace(t_map *map);
-void init_jump(t_img *img);
-void choose_good_fractale(int choice, t_map *map);
-t_map	*create_win();
-void draw_fractal(t_map *map, double x, double y);
+void				re_trace(t_map *map);
+void				init_jump(t_img *img);
+void				choose_good_fractale(int choice, t_map *map);
+t_map				*create_win();
+void				draw_fractal(t_map *map, double x, double y);
 
 #endif
